Rename packages (me.hkgumbs.tictactoe is common directory)
- Delete commented out imports in MinimaxTest
- Fix typo in SimulationTest (simlulate -> simulate)
