+++ src/me/hkgumbs/tictactoe/test/java/BoardTest.java
@@ -1,6 +1,6 @@
-package me.hkgumbs.test.java.tictactoe;
+package me.hkgumbs.tictactoe.test.java;
 
-import me.hkgumbs.main.java.tictactoe.Board;
+import me.hkgumbs.tictactoe.main.java.Board;
 import org.junit.Before;
 import org.junit.Test;
