+++ src/me/hkgumbs/tictactoe/test/java/SimulationTest.java
@@ -1,6 +1,6 @@
-package me.hkgumbs.test.java.tictactoe;
+package me.hkgumbs.tictactoe.test.java;
 
-import me.hkgumbs.main.java.tictactoe.Simulation;
+import me.hkgumbs.tictactoe.main.java.Simulation;
 import org.junit.Test;
 
 import java.io.ByteArrayInputStream;
@@ -10,7 +10,7 @@ import static org.junit.Assert.assertEquals;
 
 public class SimulationTest {
 
-    private Simulation simlulate(String contents) {
+    private Simulation simulate(String contents) {
         return new Simulation(
                 new ByteArrayInputStream(contents.getBytes()),
                 new ByteArrayOutputStream());
@@ -18,25 +18,25 @@ public class SimulationTest {
 
     @Test
     public void initialState() {
-        assertEquals(Simulation.State.INITIAL, simlulate("").state());
+        assertEquals(Simulation.State.INITIAL, simulate("").state());
     }
 
     @Test
     public void whoGoesFirstState() {
-        assertEquals(Simulation.State.HUMAN_TURN, simlulate("y\n").next());
-        assertEquals(Simulation.State.CPU_TURN, simlulate("n\n").next());
+        assertEquals(Simulation.State.HUMAN_TURN, simulate("y\n").next());
+        assertEquals(Simulation.State.CPU_TURN, simulate("n\n").next());
     }
 
     @Test
     public void humanMovesThenCpuMoves() {
-        Simulation simulation = simlulate("y\n0\n");
+        Simulation simulation = simulate("y\n0\n");
         assertEquals(Simulation.State.HUMAN_TURN, simulation.next());
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
     }
 
     @Test
     public void cpuMovesThenHumanMovesThenCpuMoves() {
-        Simulation simulation = simlulate("n\n1\n");
+        Simulation simulation = simulate("n\n1\n");
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
         assertEquals(Simulation.State.HUMAN_TURN, simulation.next());
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
@@ -44,7 +44,7 @@ public class SimulationTest {
 
     @Test
     public void simulatedGameCompletes() {
-        Simulation simulation = simlulate("n\n1\n4\n");
+        Simulation simulation = simulate("n\n1\n4\n");
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
         assertEquals(Simulation.State.HUMAN_TURN, simulation.next());
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
@@ -55,7 +55,7 @@ public class SimulationTest {
 
     @Test
     public void replayGame() {
-        Simulation simulation = simlulate("n\n1\n4\ny\n");
+        Simulation simulation = simulate("n\n1\n4\ny\n");
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
         assertEquals(Simulation.State.HUMAN_TURN, simulation.next());
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
@@ -67,7 +67,7 @@ public class SimulationTest {
 
     @Test
     public void declineReplayGame() {
-        Simulation simulation = simlulate("n\n1\n4\nn\n");
+        Simulation simulation = simulate("n\n1\n4\nn\n");
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
         assertEquals(Simulation.State.HUMAN_TURN, simulation.next());
         assertEquals(Simulation.State.CPU_TURN, simulation.next());
